requirements.txt
requests>=2.31.0
python-dateutil>=2.8.2

ics>=0.7.2
inquirer>=3.1.3
